Email: Viewing the attachments of the sent emails
The cachedFileUri is a URI of EmailProvider and not allowed to
access from other APP for the security reason. To view the
attachment, a shareable uri is needed.

Use the authority of AttachmentProvider instead of EmaiProvider to
create the cached file URI. Add the new cachedFileUri to the
AttachmentProvider, and add the query(), getType() and openfile()
operation for it.

CR-fixes: 655351

Change-Id: Ib0f990a4d14bc272c278334591e595e312ed77fb